Đang xuất , hiển thị danh sách đơn hàng ,hiển thị số lượng hàng trong giỏ hàng     =>>>>>>>>>>>>>>>>>> form quanlytaikhoan.php bị lỗi khi hiển thị cùng lúc 6- 7 đơn hàng

+++ admin/brand/admin_manage_brand.php
@@ -9,9 +9,6 @@ include_once '../../PRJ_Library/connect_DB.php';
 
 $query = "SELECT * FROM `brand`";
 $result = mysqli_query($link, $query);
-if (mysqli_num_rows($result) == 0) {
-    die("No Data");
-}
 ?>
 <html>
     <head>
@@ -20,40 +17,63 @@ if (mysqli_num_rows($result) == 0) {
     </head>
     <body>
 
-        <h2>Manage Brand</h2><hr/>
-        <p>
-        <form>
-            <p><input name="bt_add" type="submit" value="Add Brand"></p>
+        <form method="get">
+            <h2>Invoice List</h2>
+            <button name="bt_log_out">Log Out</button>
+            <hr/>   
         </form>
-
-    </p>
-
-    <table border="2">
-        <tr>
-            <th>Brand Name</th>
-            <th>Logo</th>
-            <th colspan="2">....</th>
-        </tr>
-<?php
-while ($row = mysqli_fetch_array($result)) {
-    echo "<tr>";
-    echo "<td>$row[0]</td>";
-    echo "<td>$row[1]</td>";
-    echo "<td><a href='admin_edit_brand.php?id=$row[0]'>Edit</a></td>";
-    echo "<td><a href='admin_delete_brand.php?id=$row[0]'onclick=\"javascript: return confirm('Are you sure?');\">Delete</a></td>";
-    echo "</tr>";
-}
-
-?>
-        
-    </table>
         <?php
-        if (isset($_GET["bt_add"])) {
-            header("location:admin_add_brand.php");
+        if (isset($_GET["bt_log_out"])) {
+            unset($_SESSION["admin"]);
+            header("location:../admin_log_in.php");
+            mysqli_close($link);
             exit();
         }
-      
         ?>
-    
-</body>
+
+        <form
+            <p><input name="back_product_list" type="submit" value="Back To Product List">
+                <input name="bt_add" type="submit" value="Add Brand"></p>
+        </form>
+<?php
+if (isset($_GET["bt_add"])) {
+    header("location:admin_add_brand.php");
+    mysqli_close($link);
+    exit();
+}
+if (isset($_GET["back_product_list"])) {
+    header("location:../product/admin_manage_product.php");
+    mysqli_close($link);
+    exit();
+}
+?>
+        <table border="2">
+            <tr>
+                <th>Brand Name</th>
+                <th>Logo</th>
+                <th colspan="2">....</th>
+            </tr>
+<?php
+if (mysqli_num_rows($result) == 0) {
+    echo "<tr><td><h3>No Data</h3</td></tr>";
+    mysqli_close($link);
+    exit();
+}
+?>
+            <?php
+            while ($row = mysqli_fetch_array($result)) {
+                echo "<tr>";
+                echo "<td>$row[0]</td>";
+                echo "<td>$row[1]</td>";
+                echo "<td><a href='admin_edit_brand.php?id=$row[0]'>Edit</a></td>";
+                echo "<td><a href='admin_delete_brand.php?id=$row[0]'onclick=\"javascript: return confirm('Are you sure?');\">Delete</a></td>";
+                echo "</tr>";
+            }
+            mysqli_close($link);
+            ?>
+
+        </table>
+
+
+    </body>
 </html>
